requirements.txt
# 【推荐】本地文件监听模式 - 无外部依赖
# 只需安装以下基础包，即可处理本地图片和文本文件

# 数据处理
pandas>=2.0
numpy>=1.20

# Excel 操作
openpyxl>=3.0

# 配置管理
pyyaml>=6.0

# 日志记录
loguru>=0.7

# 大模型 API
openai>=1.0

# 图像处理
opencv-python>=4.5
Pillow>=10.0

# 其他工具
requests>=2.28
python-dotenv>=1.0

# 【可选】微信直连模式 - 仅当需要直接监听 WeChat 时
# 取消下行注释可启用微信直连模式（需要 Windows 环境）
# wxauto>=3.0
